Changed the way transmission works
There is no longer an alignment control message to send to
the rocket. The system now uses checksums to detect where
each packet starts.

The program now uses threads to write to files, in order to
avoid data loss on the buffer due to the program being stuck
on I/O.